Add exporting separate functions from client-v2/api to test tree-shaking
