+++ packages/client/src/client-v2/api/getAssetsBetweenChains.ts
@@ -1,4 +1,4 @@
-import { api } from "./generateApi";
+import { api } from "../generateApi";
 
 export const assetsBetweenChains = api({
   methodName: "fungibleAssetsBetweenChainsCreate",
