+++ packages/client/src/client-v2/api/getBalances.ts
@@ -1,4 +1,4 @@
-import { api } from "./generateApi";
+import { api } from "../generateApi";
 
 export const balances = api({
   methodName: "getBalances",
