+++ packages/client/src/client-v2/api/getBridges.ts
@@ -1,4 +1,4 @@
-import { api } from "./generateApi";
+import { api } from "../generateApi";
 
 export const bridges = api({
   methodName: "getBridges",
